requirements.txt
streamlit
googlemaps
google-generativeai
google-search-results
pandas
plotly
watchdog
python-dotenv
